requirements.txt
# Multi-Tenant AI Platform - Requirements

# Core Framework
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.20
pydantic==2.12.3
pydantic-settings==2.11.0

# AI & LangChain
langchain==1.0.7
langchain-classic==1.0.0
langchain-openai==1.0.3
langchain-google-vertexai==3.0.3
langchain-core==1.0.5
langsmith==0.4.43
openai==2.8.0

# Google Cloud
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.26.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.33.0
google-cloud-tasks>=2.16.0
google-auth==2.41.1

# Database
supabase>=2.9.0
postgrest>=0.17.0
storage3>=0.8.0
supafunc>=0.6.0

# Configuration & Validation
pyyaml==6.0.1
jsonschema==4.23.0
python-dotenv==1.0.0

# Templates
jinja2==3.1.4

# PDF & Document Processing
weasyprint==66.0
pdfplumber==0.11.0
pypdf==5.1.0
pdf2image==1.17.0
pillow==12.0.0

# Excel/Data
pandas>=2.0.0
openpyxl==3.1.2
pyarrow==21.0.0

# Email
python-http-client==3.3.7
sendgrid==6.12.5

# Utilities
requests==2.32.5
python-dateutil==2.9.0.post0
pytz==2024.1
validators==0.35.0
tenacity==8.5.0

# Testing (Development)
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code Quality (Development)
black==24.8.0
flake8==7.1.1
mypy==1.11.2
